+++ manifests/dotnet.yml
@@ -210,6 +210,8 @@ tests/:
       Test_Debugger_Method_Probe_Snaphots: v2.33.0
       Test_Debugger_Mix_Log_Probe: v2.33.0
       Test_Debugger_Probe_Statuses: v2.33.0
+    test_debugger_pii.py:
+      Test_Debugger_PII_Redaction: v2.33.0
   integrations/:
     crossed_integrations/:
       test_kafka.py:
