+++ manifests/java.yml
@@ -872,6 +872,11 @@ tests/:
         '*': missing_feature
         spring-boot: v0.1  # real version not known
         uds-spring-boot: v0.1  # real version not known
+    test_debugger_pii.py:
+      Test_Debugger_PII_Redaction:         
+        '*': missing_feature
+        spring-boot: v0.1  # real version not known
+        uds-spring-boot: v0.1  # real version not known
   integrations/:
     crossed_integrations/:
       test_kafka.py:
