+++ scenario_groups.yml
@@ -72,3 +72,4 @@ DEBUGGER_SCENARIOS:
   - DEBUGGER_METHOD_PROBES_SNAPSHOT
   - DEBUGGER_LINE_PROBES_SNAPSHOT
   - DEBUGGER_MIX_LOG_PROBE
+  - DEBUGGER_PII_REDACTION
