+++ utils/proxy/core.py
@@ -229,6 +229,7 @@ class _RequestLogger:
                     "DEBUGGER_LINE_PROBES_SNAPSHOT",
                     "DEBUGGER_METHOD_PROBES_SNAPSHOT",
                     "DEBUGGER_MIX_LOG_PROBE",
+                    "DEBUGGER_PII_REDACTION",
                 ):
                     response = rc_debugger.create_rcm_probe_response(
                         request_content["client"]["client_tracer"]["language"],
