+++ utils/proxy/rc_mocked_responses_debugger_pii_redaction.json
@@ -1,0 +1,17 @@
+[
+    null,
+    [
+        {
+            "language": "",
+            "id": "log170aa-acda-4453-9111-1478a6method",
+            "where": {
+              "typeName": "ACTUAL_TYPE_NAME",
+              "methodName": "Pii",
+              "sourceFile": null
+            },
+            "evaluateAt": "EXIT",
+            "captureSnapshot": true,
+            "maxFieldCount" : 500
+          }
+    ]
+]
